Crate libp2p_core

Transports, upgrades, multiplexing and node handling of libp2p.

The main concepts of libp2p-core are:

A PeerId is a unique global identifier for a node on the network. Each node must have a different PeerId. Normally, a PeerId is the hash of the public key used to negotiate encryption on the communication channel, thereby guaranteeing that they cannot be spoofed.

The Transport trait defines how to reach a remote node or listen for incoming remote connections. See the transport module.

The StreamMuxer trait is implemented on structs that hold a connection to a remote and can subdivide this connection into multiple substreams. See the muxing module.

The UpgradeInfo, InboundUpgrade and OutboundUpgrade traits define how to upgrade each individual substream to use a protocol. See the upgrade module.

Re-exports

pub use multiaddr;

pub use muxing::StreamMuxer;

pub use identity::PublicKey;

pub use transport::Transport;

pub use upgrade::InboundUpgrade;

pub use upgrade::OutboundUpgrade;

pub use upgrade::UpgradeInfo;

pub use upgrade::ProtocolName;

Modules

either

identity
A node's network identity keys.

muxing
Muxing is the process of splitting a connection into multiple substreams.

nodes
Low-level networking primitives.

transport
Connection-oriented communication channels.

upgrade
Contains everything related to upgrading a connection or a substream to use a protocol.

Structs

Multiaddr
Representation of a Multiaddr.

Negotiated
A stream after it has been negotiated.

PeerId
Identifier of a peer of the network.

Enums

ConnectedPoint
How we connected to a node.

Endpoint

UpgradeError
Error that can happen when upgrading a connection or substream to use a protocol.

Functions

address_translation
Perform IP address translation.