Module libp2p_core::identity

A node’s network identity keys.

Such identity keys can be randomly generated on every startup, but using already existing, fixed keys is usually required. Though libp2p uses other crates (e.g. ed25519_dalek) internally, such details are not exposed as part of libp2p’s public interface to keep them easily upgradable or replaceable (e.g. to ed25519_zebra). Consequently, keys of external ed25519 or secp256k1 crates cannot be directly converted into libp2p network identities. Instead, loading fixed keys must use the standard, thus more portable binary representation of the specific key type (e.g. ed25519 binary format). All key types have functions to enable conversion to/from their binary representations.

Modules

ed25519

Ed25519 keys.

error

Errors during identity key operations.

rsa

RSA keys.

secp256k1

Secp256k1 keys.

Enums

Keypair

Identity keypair of a node.

PublicKey

The public key of a node’s identity keypair.