Module libp2p_core::nodes::network

Structs

IncomingConnectionEvent
A new connection arrived on a listener.

IncomingInfo
Information about an incoming connection currently being negotiated.

Network
Implementation of Stream that handles the nodes.

PeerConnected
Access to a peer we are connected to.

PeerNotConnected
Access to a peer we're not connected to.

PeerPendingConnect
Access to a peer we are attempting to connect to.

Enums

IncomingError
Error that can happen on an incoming connection.

NetworkEvent
Event that can happen on the Network.

NetworkReachError
Error that can happen when trying to reach a node.

Peer
State of a peer in the system.

PeerPotentialConnect
Peer we are potentially going to connect to.

PeerState
State of a peer.

StartTakeOver
Result of TaskEntry::start_take_over.

UnknownPeerDialErr
Error that can happen when dialing a node with an unknown peer ID.