Enum libp2p_core::swarm::NetworkBehaviourAction

pub enum NetworkBehaviourAction<TInEvent, TOutEvent> {
    GenerateEvent(TOutEvent),
    DialAddress {
        address: Multiaddr,
    },
    DialPeer {
        peer_id: PeerId,
    },
    SendEvent {
        peer_id: PeerId,
        event: TInEvent,
    },
    ReportObservedAddr {
        address: Multiaddr,
    },
}

An action that a NetworkBehaviour can trigger in the [Swarm] in whose context it is executing.

Variants
GenerateEvent(TOutEvent)

Instructs the Swarm to return an event when it is being polled.

DialAddress

Instructs the swarm to dial the given multiaddress, with no knowledge of the PeerId that may be reached.

Fields of DialAddress

address: Multiaddr

The address to dial.

DialPeer

Instructs the swarm to dial a known PeerId.

The addresses_of_peer method is called to determine which addresses to attempt to reach.

If we were already trying to dial this node, the addresses that are not yet in the queue of addresses to try are added back to this queue.

On success, NetworkBehaviour::inject_connected is invoked. On failure, NetworkBehaviour::inject_dial_failure is invoked.

Fields of DialPeer

peer_id: PeerId

The peer to try reach.

SendEvent

Instructs the Swarm to send a message to the handler dedicated to the connection with the peer.

If the Swarm is connected to the peer, the message is delivered to the remote's protocol handler. If there is no connection to the peer, the message is ignored. To ensure delivery, the NetworkBehaviour must keep track of connected peers.

Note that even if the peer is currently connected, connections can get closed at any time and thus the message may not reach its destination.

Fields of SendEvent

peer_id: PeerId

The peer to which to send the message.

event: TInEvent

The message to send.

ReportObservedAddr

Informs the Swarm about a multi-address observed by a remote for the local node.

It is advisable to issue ReportObservedAddr actions at a fixed frequency per node. This way address information will be more accurate over time and individual outliers carry less weight.

Fields of ReportObservedAddr

address: Multiaddr

The observed address of the local node.