libp2p_core
Module transport

Connection-oriented communication channels.

The main entity of this module is the Transport trait, which provides an interface for establishing connections with other nodes, thereby negotiating any desired protocols. The rest of the module defines combinators for modifying a transport through composition with other transports or protocol upgrades.

Re-exports

pub use self::choice::OrTransport;

pub use self::memory::MemoryTransport;

pub use self::upgrade::Upgrade;

Modules

and_then

choice

dummy

global_only

map

map_err

memory

timeout

Transports with timeouts on the connection setup.

upgrade

Configuration of transport protocol upgrades.

Structs

Boxed

A Boxed transport is a Transport whose Dial, Listener and ListenerUpgrade futures are Boxed and only the Output type is captured in a type variable.

ListenerId

The ID of a single listener.

OptionalTransport

Transport that is possibly disabled.

Enums

TransportError

An error during dialing or listening on a Transport.

TransportEvent

Event produced by Transports.

Traits

Transport

A transport provides connection-oriented communication between two peers through ordered streams of data (i.e. connections).