Struct libp2p_core::transport::upgrade::Builder

pub struct Builder<T> { /* private fields */ }

A Builder facilitates upgrading of a Transport for use with a Swarm.

The upgrade process is defined by the following stages:

authenticate{1} -> apply{*} -> multiplex{1}

It thus enforces the following invariants on every transport obtained from multiplex:

The transport must be authenticated and multiplexed.

Authentication must precede the negotiation of a multiplexer.

Applying a multiplexer is the last step in the upgrade process.

The Transport::Output conforms to the requirements of a Swarm, namely a tuple of a PeerId (from the authentication upgrade) and a StreamMuxer (from the multiplexing upgrade).

Implementations

impl<T> Builder<T>
where T: Transport, T::Error: 'static,

pub fn new(inner: T, version: Version) -> Builder<T>

Creates a Builder over the given (base) Transport.

pub fn authenticate<C, D, U, E>( self, upgrade: U ) -> Authenticated<AndThen<T, impl FnOnce(C, ConnectedPoint) -> Authenticate<C, U> + Clone>>
where T: Transport<Output = C>, C: AsyncRead + AsyncWrite + Unpin, D: AsyncRead + AsyncWrite + Unpin, U: InboundUpgrade<Negotiated<C>, Output = (PeerId, D), Error = E> + OutboundUpgrade<Negotiated<C>, Output = (PeerId, D), Error = E> + Clone, E: Error + 'static,

Upgrades the transport to perform authentication of the remote.

The supplied upgrade receives the I/O resource C and must produce a pair (PeerId, D), where D is a new I/O resource. The upgrade must thus at a minimum identify the remote, which typically involves the use of a cryptographic authentication protocol in the context of establishing a secure channel.

Transitions

I/O upgrade: C -> (PeerId, D).

Transport output: C -> (PeerId, D)