Struct libp2p_core::transport::upgrade::Multiplexed

pub struct Multiplexed<T>(_);

A authenticated and multiplexed transport, obtained from Authenticated::multiplex.

Implementations

impl<T> Multiplexed<T>

pub fn boxed<M>(self) -> Boxed<(PeerId, StreamMuxerBox)> where
T: Transport<Output = (PeerId, M)> + Sized + Clone + Send + Sync + 'static,
T::Dial: Send + 'static,
T::Listener: Send + 'static,
T::ListenerUpgrade: Send + 'static,
T::Error: Send + Sync,
M: StreamMuxer + Send + Sync + 'static,
M::Substream: Send + 'static,
M::OutboundSubstream: Send + 'static,

Boxes the authenticated, multiplexed transport, including the StreamMuxer and custom transport errors.

pub fn timeout(self, timeout: Duration) -> Multiplexed<TransportTimeout<T>>

Adds a timeout to the setup and protocol upgrade process for all inbound and outbound connections established through the transport.

pub fn outbound_timeout(
self,
timeout: Duration
) -> Multiplexed<TransportTimeout<T>>

Adds a timeout to the setup and protocol upgrade process for all outbound connections established through the transport.

pub fn inbound_timeout(
self,
timeout: Duration
) -> Multiplexed<TransportTimeout<T>>

Adds a timeout to the setup and protocol upgrade process for all inbound connections established through the transport.