Struct libp2p_core::upgrade::MapOutboundUpgrade

pub struct MapOutboundUpgrade<U, F> { /* private fields */ }

๐Deprecated: Deprecated without replacement because it is not used within rust-libp2p.

Wraps around an upgrade and applies a closure to the output.

Implementations

impl<U, F> MapOutboundUpgrade<U, F>

pub fn new(upgrade: U, fun: F) -> Self