Trait libp2p_core::upgrade::OutboundUpgradeExt

pub trait OutboundUpgradeExt<C>: OutboundUpgrade<C> {
    // Provided methods
    fn map_outbound<F, T>(self, f: F) -> MapOutboundUpgrade<Self, F>
       where Self: Sized,
             F: FnOnce(Self::Output) -> T { ... }
    fn map_outbound_err<F, T>(self, f: F) -> MapOutboundUpgradeErr<Self, F>
       where Self: Sized,
             F: FnOnce(Self::Error) -> T { ... }
}

👎Deprecated: Will be removed without replacement because it is not used within rust-libp2p.

Extention trait for OutboundUpgrade. Automatically implemented on all types that implement OutboundUpgrade.

Provided Methods

fn map_outbound<F, T>(self, f: F) -> MapOutboundUpgrade<Self, F>where Self: Sized, F: FnOnce(Self::Output) -> T,

👎Deprecated: Will be removed without replacement because it is not used within rust-libp2p.

Returns a new object that wraps around Self and applies a closure to the Output.

fn map_outbound_err<F, T>(self, f: F) -> MapOutboundUpgradeErr<Self, F>where Self: Sized, F: FnOnce(Self::Error) -> T,

👎Deprecated: Will be removed without replacement because it is not used within rust-libp2p.

Returns a new object that wraps around Self and applies a closure to the Error.