Crate libp2p_core

Transports, upgrades, multiplexing and node handling of libp2p.

The main concepts of libp2p-core are:

The Transport trait defines how to reach a remote node or listen for incoming remote connections. See the transport module.

The StreamMuxer trait is implemented on structs that hold a connection to a remote and can subdivide this connection into multiple substreams. See the muxing module.

The UpgradeInfo, InboundUpgrade and OutboundUpgrade traits define how to upgrade each individual substream to use a protocol. See the upgrade module.

Re-exports

pub use connection::ConnectedPoint;

pub use connection::Endpoint;

pub use muxing::StreamMuxer;

pub use peer_record::PeerRecord;

pub use signed_envelope::SignedEnvelope;

pub use transport::Transport;

pub use upgrade::InboundUpgrade;

pub use upgrade::OutboundUpgrade;

pub use upgrade::UpgradeInfo;

pub use multiaddr;

pub use multihash;

Modules

connection

either

muxing

Muxing is the process of splitting a connection into multiple substreams.

peer_record

signed_envelope

transport

Connection-oriented communication channels.

upgrade

Contains everything related to upgrading a connection or a substream to use a protocol.

Structs

DecodeError

Multiaddr

Representation of a Multiaddr.

Functions

address_translation

Perform IP address translation.

Type Aliases

Negotiated