Enum libp2p_core::network::NetworkEvent

pub enum NetworkEvent<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId> where
    TTrans: Transport,
    THandler: IntoConnectionHandler<TConnInfo>,  {
    ListenerClosed {
        listener_id: ListenerId,
        addresses: Vec<Multiaddr>,
        reason: Result<(), TTrans::Error>,
    },
    ListenerError {
        listener_id: ListenerId,
        error: TTrans::Error,
    },
    NewListenerAddress {
        listener_id: ListenerId,
        listen_addr: Multiaddr,
    },
    ExpiredListenerAddress {
        listener_id: ListenerId,
        listen_addr: Multiaddr,
    },
    IncomingConnection(IncomingConnectionEvent<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>),
    IncomingConnectionError {
        local_addr: Multiaddr,
        send_back_addr: Multiaddr,
        error: PendingConnectionError<TTrans::Error>,
    },
    ConnectionEstablished {
        connection: EstablishedConnection<'a, TInEvent, TConnInfo, TPeerId>,
        num_established: NonZeroU32,
    },
    ConnectionError {
        id: ConnectionId,
        connected: Connected<TConnInfo>,
        error: ConnectionError<<THandler::Handler as ConnectionHandler>::Error>,
        num_established: u32,
    },
    DialError {
        attempts_remaining: u32,
        peer_id: TPeerId,
        multiaddr: Multiaddr,
        error: PendingConnectionError<TTrans::Error>,
    },
    UnknownPeerDialError {
        multiaddr: Multiaddr,
        error: PendingConnectionError<TTrans::Error>,
    },
    ConnectionEvent {
        connection: EstablishedConnection<'a, TInEvent, TConnInfo, TPeerId>,
        event: TOutEvent,
    },
}

Event that can happen on the Network.

Variants

ListenerClosed

One of the listeners gracefully closed.

Fields of ListenerClosed

listener_id: ListenerId

The listener ID that closed.

addresses: Vec<Multiaddr>

The addresses that the listener was listening on.

reason: Result<(), TTrans::Error>

Reason for the closure. Contains Ok(()) if the stream produced None, or Err if the stream produced an error.

ListenerError

One of the listeners reported a non-fatal error.

Fields of ListenerError

listener_id: ListenerId

The listener that errored.

error: TTrans::Error

The listener error.

NewListenerAddress

One of the listeners is now listening on an additional address.

Fields of NewListenerAddress

listener_id: ListenerId

The listener that is listening on the new address.

listen_addr: Multiaddr

The new address the listener is now also listening on.

ExpiredListenerAddress

One of the listeners is no longer listening on some address.

Fields of ExpiredListenerAddress

listener_id: ListenerId

The listener that is no longer listening on some address.

listen_addr: Multiaddr

The expired address.

IncomingConnection(IncomingConnectionEvent<'a, TTrans, TInEvent, TOutEvent, THandler, TConnInfo, TPeerId>)

A new connection arrived on a listener.

IncomingConnectionError

An error happened on a connection during its initial handshake.

This can include, for example, an error during the handshake of the encryption layer, or the connection unexpectedly closed.

Fields of IncomingConnectionError

local_addr: Multiaddr

Local connection address.

send_back_addr: Multiaddr

Address used to send back data to the remote.

error: PendingConnectionError<TTrans::Error>

The error that happened.

ConnectionEstablished

A new connection to a peer has been opened.

Fields of ConnectionEstablished

connection: EstablishedConnection<'a, TInEvent, TConnInfo, TPeerId>

The newly established connection.

num_established: NonZeroU32

The total number of established connections to the same peer, including the one that has just been opened.

ConnectionError

An established connection to a peer has encountered an error.

The connection is closed as a result of the error.

Fields of ConnectionError

id: ConnectionId

The ID of the connection that encountered an error.

connected: Connected<TConnInfo>

Information about the connection that encountered the error.

error: ConnectionError<<THandler::Handler as ConnectionHandler>::Error>

The error that occurred.

num_established: u32

The remaining number of established connections to the same peer.

DialError

A dialing attempt to an address of a peer failed.

Fields of DialError

attempts_remaining: u32

The number of remaining dialing attempts.

peer_id: TPeerId

Id of the peer we were trying to dial.

multiaddr: Multiaddr

The multiaddr we failed to reach.

error: PendingConnectionError<TTrans::Error>

The error that happened.

UnknownPeerDialError

Failed to reach a peer that we were trying to dial.

Fields of UnknownPeerDialError

multiaddr: Multiaddr

The multiaddr we failed to reach.

error: PendingConnectionError<TTrans::Error>

The error that happened.

ConnectionEvent

An established connection produced an event.

Fields of ConnectionEvent

connection: EstablishedConnection<'a, TInEvent, TConnInfo, TPeerId>

The connection on which the event occurred.

event: TOutEvent

Event that was produced by the node.