Module libp2p_core::network::peer

Structs

ConnectedPeer

The representation of a peer in a Network to whom at least one established connection exists. There may also be additional ongoing dialing attempts to the peer.

DialingAttempt

A DialingAttempt is a pending outgoing connection attempt to a known / expected remote peer ID.

DialingAttemptIter

An iterator over the ongoing dialing attempts to a peer.

DialingPeer

The representation of a peer in a Network to whom a dialing attempt is ongoing. There may already exist other established connections to this peer.

DisconnectedPeer

The representation of a peer to whom the Network has currently neither an established connection, nor an ongoing dialing attempt initiated by the local peer.

Enums

Peer

The possible representations of a peer in a Network, as seen by the local node.