Module libp2p_core::nodes

Low-level networking primitives.

Contains structs that are aiming at providing very precise control over what happens over the network.

The more complete and highest-level struct is the RawSwarm. The RawSwarm directly or indirectly uses all the other structs of this module.

Re-exports

pub use self::collection::ConnectionInfo;

pub use self::node::Substream;

pub use self::handled_node::NodeHandlerEvent;

pub use self::handled_node::NodeHandlerEndpoint;

pub use self::raw_swarm::ConnectedPoint;

pub use self::raw_swarm::Peer;

pub use self::raw_swarm::RawSwarm;

pub use self::raw_swarm::RawSwarmEvent;

Modules

collection

handled_node

handled_node_tasks

listeners
Manage listening on multiple multiaddresses at once.

node

raw_swarm