Trait libp2p_core::upgrade::InboundUpgrade

pub trait InboundUpgrade<C>: UpgradeInfo {
    type Output;
    type Error;
    type Future: Future<Item = Self::Output, Error = Self::Error>;
    fn upgrade_inbound(
        self, 
        socket: Negotiated<C>, 
        info: Self::Info
    ) -> Self::Future;
}

Possible upgrade on an inbound connection or substream.

Associated Types

type Output

Output after the upgrade has been successfully negotiated and the handshake performed.

type Error

Possible error during the handshake.

type Future: Future<Item = Self::Output, Error = Self::Error>

Future that performs the handshake with the remote.

Loading content...
Required methods

fn upgrade_inbound(
self,
socket: Negotiated<C>,
info: Self::Info
) -> Self::Future

After we have determined that the remote supports one of the protocols we support, this method is called to start the handshake.

The info is the identifier of the protocol, as produced by protocol_info.

Loading content...